add a new field to represent schemes after refs are resolved
I need to be able to validate the data before and after the refs are
resolved. So, I've implemented a new property that is the json schema
after refs are resolved. But, I think that I should not do two schemes,
but rather one schema before refs are resolved, and a custom schema made
on the fly by the config manager. This schema would be made on the fly,
as refs get resolved.